test: make fake mapper globally available
